+++ AuctionMarket/src/main/java/com/portfolio/auctionmarket/domain/bids/entity/Bid.java
@@ -31,6 +31,6 @@ public class Bid extends BaseCreatedAt {
     private User bidder;
 
     @Column(name = "bid_price", nullable = false)
-    private Integer bidPrice;
+    private Long bidPrice;
 
 }
